Split components out of utils.jsx and rename files
